Factor timer code into separate driver files
