+++ src/systimer_null.c
@@ -1,0 +1,20 @@
+int SyHaveAlarms = 0;
+
+/* stub implementations */
+
+static void SyInitAlarm( void )
+{
+    /* No initialisation in this case */
+    return;
+}
+
+void SyInstallAlarm ( UInt seconds, UInt nanoseconds )
+{
+    assert(0);
+    return;
+}
+
+void SyStopAlarm(UInt *seconds, UInt *nanoseconds) {
+    assert(0);
+    return;
+}
